services/api_service/requirements.txt
django
djangorestframework
python-dotenv
google-cloud-pubsub
